Update builder-protocol compilation with new syntax.
The protoc-rust updated the syntax used in build.rs for code generation, deprecating the syntax we were previously using.
This commit updates us to the new `Codegen` syntax.

In addition, the components/libbuild-protocols.rs was only used in components/builder-protocols/build.rs, so those two files
have been combined, as well as refactored to Rust 2018 syntax for modules.
